nft minter/server/requirements.txt
flask
flask-cors
torch
diffusers
transformers
requests
python-dotenv
